+++ fiddles/java/fiddle-0003-SpringFox/springfox/boot-static-docs/src/main/java/springfoxdemo/staticdocs/SwaggerConfig.java
@@ -1,0 +1,11 @@
+package springfoxdemo.staticdocs;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@EnableSwagger2
+public class SwaggerConfig {
+
+}